requirements.txt
blinker==1.4
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
Flask==2.1.2
Flask-Mail==0.9.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
py-dotenv==0.1
pymongo==4.1.1
pypiwin32==223
python-dotenv==0.20.0
pywin32==304
Werkzeug==2.1.2
